Remove core and create common for factoring
