--- services/dauth-service/src/core/mod.rs
@@ -1,3 +1,0 @@
-pub mod auth_vectors;
-pub mod confirm_keys;
-pub mod users;
